backend/core_app/permissions.py
from rest_framework.permissions import BasePermission, SAFE_METHODS


def is_admin_user(user):
    """Return True if user is authenticated and has admin privileges."""
    if not user or not user.is_authenticated:
        return False
    return bool(
        getattr(user, 'is_superuser', False)
        or getattr(user, 'is_staff', False)
        or getattr(user, 'role', None) == 'admin'
    )


class IsAdminRole(BasePermission):
    def has_permission(self, request, view):
        return is_admin_user(request.user)


class IsAdminOrReadOnly(BasePermission):
    def has_permission(self, request, view):
        if request.method in SAFE_METHODS:
            return True
        return is_admin_user(request.user)
