backend/requirements.txt
Django>=4.2,<4.3

djangorestframework>=3.14,<3.16

djangorestframework-simplejwt>=5.3,<5.4

django-cors-headers>=4.3,<4.5

python-decouple>=3.8,<3.9

pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=4.1,<5.0
coverage>=7.4,<8.0
